requirements.txt
pyspark==3.5.0
requests==2.31.0
pandas==2.1.0
numpy==1.24.0
python-dotenv==1.0.0
plotly==5.17.0
dash==2.14.0
scikit-learn==1.3.0
joblib==1.3.0
